requirements.txt
livekit-agents[silero,turn-detector]~=1.4
livekit-plugins-openai~=1.4
livekit-plugins-noise-cancellation~=0.2
python-dotenv
fastapi
uvicorn[standard]
